fix: added api helper
